Add parameter clean to only delete volumes when its passed.

+++ launch-testing.sh
@@ -6,9 +6,12 @@ NOTIFY_TIMEOUT=10000
 params=${1}
 
 docker-compose rm -f -v
-sudo rm -fr volumes/config
-sudo rm -fr volumes/mysql/*
-docker-compose build ${params}
+if [ "${params}" == "clean" ]; then
+  sudo rm -fr volumes/config
+  sudo rm -fr volumes/mysql/*
+  params="--no-cache"
+fi
+#docker-compose build ${params}
 if [ $? -gt 0 ]; then
   out=$(echo ${out}|tail -n 10)
   notify-send 'App rebuild failure' "There was an error building the container, see console for build output" -t ${NOTIFY_TIMEOUT} -i dialog-error && \
